+++ type-generator/src/transformer/rename_all.rs
@@ -1,55 +1,9 @@
-use crate::ir::{RenameRule, RustEnumMemberKind, RustSegment, RustStructAttr, SerdeContainerAttr};
+use crate::{
+    case::{detect_case, CaseConvention},
+    ir::{RustEnumMemberKind, RustSegment, RustStructAttr, SerdeContainerAttr},
+};
 
 pub fn adapt_rename_all(segment: &mut RustSegment) -> Option<()> {
-    #[derive(Debug, Clone, Copy, PartialEq)]
-    enum CaseConvention {
-        Lower,
-        Upper,
-        Snake,
-        Pascal,
-        ScreamingSnake,
-    }
-    use CaseConvention::*;
-    /// expect input follows one of above case
-    fn detect_case(s: &str) -> CaseConvention {
-        if s.starts_with(char::is_uppercase) {
-            if s.contains('_') {
-                ScreamingSnake
-            } else if s.chars().all(char::is_uppercase) {
-                Upper
-            } else {
-                Pascal
-            }
-        } else if s.contains('_') {
-            Snake
-        } else {
-            Lower
-        }
-    }
-    impl CaseConvention {
-        fn cast(&mut self, other: Self) -> Option<()> {
-            if self == &other {
-                return Some(());
-            }
-            match (&self, &other) {
-                (Lower, Snake) | (Upper, ScreamingSnake) => {
-                    *self = other;
-                }
-                (Snake, Lower) | (ScreamingSnake, Upper) => {
-                    // do nothing
-                }
-                _ => None?,
-            }
-            Some(())
-        }
-        fn into_rename_rule(self) -> RenameRule {
-            match self {
-                Lower | Snake => RenameRule::SnakeCase,
-                Pascal => RenameRule::PascalCase,
-                Upper | ScreamingSnake => RenameRule::ScreamingSnakeCase,
-            }
-        }
-    }
     if let RustSegment::Enum(re) = segment {
         let mut conv: Option<CaseConvention> = None;
         for memb in &re.member {
